requirements.txt
yfinance==0.2.28
openai==1.35.0
python-dotenv==1.0.0
pandas==2.0.3
requests==2.31.0